[move] repository 패키지 생성 및 경로 수정 #22
